MXDataHandler does not implement IMXEventListener (it was useless) MxEventDispatcher is created to dispatch event to the listeners
